restructure & rename stuff
