+++ contracts/starknet/interfaces/i_execution_strategy.cairo
@@ -4,7 +4,7 @@ from contracts.starknet.lib.eth_address import EthAddress
 from starkware.cairo.common.uint256 import Uint256
 
 @contract_interface
-namespace IExecutionStrategy:
+namespace i_execution_strategy:
     func execute(
         proposal_outcome : felt,
         execution_hash : Uint256,
