+++ contracts/starknet/interfaces/i_voting_strategy.cairo
@@ -4,7 +4,7 @@ from contracts.starknet.lib.eth_address import EthAddress
 from starkware.cairo.common.uint256 import Uint256
 
 @contract_interface
-namespace IVotingStrategy:
+namespace i_voting_strategy:
     func get_voting_power(
         timestamp : felt, address : EthAddress, params_len : felt, params : felt*
     ) -> (voting_power : Uint256):
