+++ contracts/starknet/voting_strategies/vanilla.cairo
