+++ package.json
@@ -15,7 +15,7 @@
     "test:crosschain:ci": "concurrently './scripts/chain.sh' './scripts/test-crosschain.sh'",
     "chain:l1": "hardhat node",
     "chain:l2": "starknet-devnet -p 8000",
-    "chain": "concurrently 'yarn chain:l1' 'yarn chain:l2'",
+    "chain": "yarn chain:l1 && yarn chain:l2",
     "format:l1": "prettier --write 'contracts/**/*.sol'",
     "check-format:l1": "prettier -c 'contracts/**/*.sol'",
     "format:l2": "cairo-format -i contracts/starknet/**/*.cairo",
