+++ test/starknet/lib_words.ts
@@ -5,7 +5,7 @@ import { wordsToUint } from './shared/helpers';
 
 async function setup() {
   const testWordsFactory = await starknet.getContractFactory(
-    './contracts/starknet/testContracts/test_words.cairo'
+    './contracts/starknet/test_contracts/test_words.cairo'
   );
   const testWords = await testWordsFactory.deploy();
   return {
